backend/requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
flask==3.1.1
flask-cors==6.0.1
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
mysql-connector-python==9.4.0
python-dotenv==1.1.1
werkzeug==3.1.3
